clean BrainfuckStream to file.
