+++ src/Brainfuck.Net/Program.cs
@@ -27,18 +27,5 @@ namespace Brainfuck.Net
             }
             
         }
-
-        public class BrainfuckStream : IBrainfuckStream
-        {
-            public void Write(char c)
-            {
-                Console.Write(c);
-            }
-
-            public char Read()
-            {
-                return (char)Console.Read();
-            }
-        }
     }
 }
